cleaned up some examples in regards to structs

+++ examples/blink_wo_delay/main.c
@@ -13,10 +13,10 @@
 // constants won't change. Used here to 
 // set pin number and time interval:
 const uint8_t ledPin =  3;          // the number of the LED pin
-const uint32_t interval = 1000;     // interval at which to blink (milliseconds)
+const uint16_t interval = 1000;     // interval at which to blink (milliseconds)
 
 // Variables will change:
-uint32_t previousMillis = 0;        // will store last time LED was updated
+uint16_t previousMillis = 0;        // will store last time LED was updated
 
 int main (void)
 {
@@ -32,7 +32,7 @@ int main (void)
         // difference between the current time and last time you blinked 
         // the LED is bigger than the interval at which you want to 
         // blink the LED.
-        uint32_t currentMillis = millis();
+        uint16_t currentMillis = millis();
 
         if(currentMillis - previousMillis > interval) {
         // save the last time you blinked the LED
